finish adventure

+++ Assets/Scripts/Adventure/Events/TwilightForestHunt.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 namespace ThisGame.Adventure.Events {
     [CreateAssetMenu(fileName = "TwilightForestHunt", menuName = "Adventure/Event/TwilightForestHunt")]
     public class TwilightForestHunt : LocationEvent {
-        public override bool IsSuccess(Dictionary<ItemDescription, uint> along, TimeWeather dtw, uint[] count) {
+        public override bool IsSuccess(Dictionary<ItemDescription, int> along, TimeWeather dtw, int[] count) {
             if(!along.Any(kvp => kvp.Key.properties[(int)ItemProperty.Metal] > 2))
                 return false;
-            count[0] = (uint)Random.Range(1, 3);
+            count[0] = Random.Range(1, 3);
             return true;
         }
     }
